POCs/underwriting-automation/requirements.txt
absl-py==0.13.0
aiofile==3.7.4
aiohttp==3.7.4.post0
anyio==3.3.4
Artisan==0.0.0
astunparse==1.6.3
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
cachetools==4.2.2
caio==0.9.5
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.5
colorama==0.4.4
cycler==0.10.0
distro==1.6.0
Extractor==0.5
flatbuffers==2.0
fuzzywuzzy==0.18.0
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.40.0
h11==0.12.0
h5py==3.6.0
httpcore==0.13.7
httpx==0.20.0
idna==3.2
iniconfig==1.1.1
install==1.3.4
jellyfish==0.6.1
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.4
matplotlib==3.4.3
mergedeep==1.3.4
multidict==5.1.0
numpy==1.21.2
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.3
Pillow==8.3.2
pluggy==1.0.0
protobuf==3.18.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMuPDF==1.19.0
pyparsing==2.4.7
pytesseract==0.3.8
pytest==6.2.5
pytest-aiohttp==0.3.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rsa==4.7.2
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
sniffio==1.2.0
tabula-py==2.3.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
thop==0.0.31.post2005241907
toml==0.10.2
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.2
typing-extensions==3.10.0.2
urllib3==1.26.6
us==2.0.2
validity==0.1.0
Werkzeug==2.0.1
wrapt==1.13.3
yarl==1.6.3
